test: consolidate remote account provider tests
